+++ demos/vue-supabase-todolist/src/plugins/powerSync.ts
@@ -1,10 +1,10 @@
 import { AppSchema } from '@/library/powersync/AppSchema';
-import { WASQLitePowerSyncDatabaseOpenFactory } from '@powersync/web';
+import { PowerSyncDatabase } from '@powersync/web';
 import { createPowerSyncPlugin } from '@powersync/vue';
 
-export const powerSync = new WASQLitePowerSyncDatabaseOpenFactory({
-  dbFilename: 'example-vue-todo.db',
+export const powerSync = new PowerSyncDatabase({
+  database: { dbFilename: 'example-vue-todo.db' },
   schema: AppSchema
-}).getInstance();
+});
 
 export const powerSyncPlugin = createPowerSyncPlugin({ database: powerSync });