+++ packages/web/src/db/adapters/wa-sqlite/WASQLitePowerSyncDatabaseOpenFactory.ts
@@ -1,11 +1,21 @@
 import { AbstractPowerSyncDatabase, DBAdapter, PowerSyncDatabaseOptions } from '@powersync/common';
 import { PowerSyncDatabase } from '../../../db/PowerSyncDatabase';
-import { WASQLiteDBAdapter } from './WASQLiteDBAdapter';
 import { AbstractWebPowerSyncDatabaseOpenFactory } from '../AbstractWebPowerSyncDatabaseOpenFactory';
+import { WASQLiteOpenFactory } from './WASQLiteOpenFactory';
 
+/**
+ * @deprecated {@link PowerSyncDatabase} can now be constructed directly
+ * @example
+ * ```typescript
+ * const powersync = new PowerSyncDatabase({database: {
+ *  dbFileName: 'powersync.db'
+ * }});
+ * ```
+ */
 export class WASQLitePowerSyncDatabaseOpenFactory extends AbstractWebPowerSyncDatabaseOpenFactory {
   protected openDB(): DBAdapter {
-    return new WASQLiteDBAdapter({ ...this.options, flags: this.resolveDBFlags() });
+    const factory = new WASQLiteOpenFactory(this.options);
+    return factory.openDB();
   }
 
   generateInstance(options: PowerSyncDatabaseOptions): AbstractPowerSyncDatabase {