+++ demos/angular-supabase-todolist/CHANGELOG.md
@@ -1,5 +1,12 @@
 # PowerSync Angular Demo
 
+## 0.0.21
+
+### Patch Changes
+
+- Updated dependencies [dca599f]
+  - @powersync/web@1.2.0
+
 ## 0.0.20
 
 ### Patch Changes